// src/models/interfaces/Color.ts
import ColorSchemaEnumInterface from './ColorSchemaEnum'

export default interface ColorInterface {
  colorSchema: ColorSchemaEnumInterface | number
  veryDarkBrown: string
  darkGray: string
  brown: string
  transparentBrown: string
  lightBrown: string
  veryLightBrown: string
  gray: string
  lightGray: string
  white: string
  black: string
  veryLightGray: string
}